[REFACTOR] Apply SRP better on change password
